@@ -18,7 +18,7 @@
         }
     ],
     "require": {
-        "php": ">=8.1",
+        "php": ">=8.2",
         "illuminate/support": "^10.0|^11.0|^12.0",
         "illuminate/cache": "^10.0|^11.0|^12.0",
         "intervention/image": "^3.4"
@@ -29,8 +29,8 @@
     },
     "require-dev": {
         "roave/security-advisories": "dev-latest",
-        "phpunit/phpunit": "^10.5",
-        "mockery/mockery": "~1.3",
+        "phpunit/phpunit": "^11.5.3",
+        "mockery/mockery": "^1.6.7",
         "php-coveralls/php-coveralls": "^2.1",
         "pestphp/pest": "^2.34|^3.7",
         "pestphp/pest-plugin-type-coverage": "^2.8|^3.3"
