@@ -4,13 +4,12 @@ namespace Laravolt\Avatar\Test;
 
 use InvalidArgumentException;
 use Mockery;
+use PHPUnit\Framework\Attributes\Test;
 use stdClass;
 
 class AvatarPhpTest extends \PHPUnit\Framework\TestCase
 {
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_override_attributes_when_instantiated()
     {
         $config = [
@@ -42,9 +41,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals(false, $avatar->getAttribute('ascii'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_override_attributes_after_set_name()
     {
         $config = ['backgrounds' => ['#000000', '#111111'], 'foregrounds' => ['#EEEEEE', '#FFFFFF']];
@@ -55,9 +52,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals('#FFFFFF', $avatar->getAttribute('foreground'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_has_correct_random_background()
     {
         $config = [
@@ -73,9 +68,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals('#111111', $avatar->getAttribute('foreground'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_has_different_random_background()
     {
         $config = [
@@ -95,9 +88,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals('#111111', $avatar2->getAttribute('background'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_resolve_random_themes_and_then_overrides()
     {
         $config = [
@@ -127,9 +118,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals('#000000', $avatar1->getAttribute('foreground'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_handle_invalid_theme()
     {
         $config = [
@@ -148,9 +137,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $avatar->setTheme(['satu', 'dua'])->buildAvatar();
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_create_initials_from_name()
     {
         $avatar = new \Laravolt\Avatar\Avatar();
@@ -160,9 +147,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals('BH', $avatar->getAttribute('initials'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_set_chars_length()
     {
         $avatar = new \Laravolt\Avatar\Avatar();
@@ -174,9 +159,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals('BHW', $avatar->getAttribute('initials'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_accept_valid_font_file()
     {
         $font = __DIR__.'/fonts/rockwell.ttf';
@@ -187,9 +170,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($font, $avatar->getAttribute('font'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_cannot_accept_invalid_font_file()
     {
         $font = __DIR__.'/fonts/invalid-font.ttf';
@@ -200,9 +181,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertNotEquals($font, $avatar->getAttribute('font'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_generate_base64()
     {
         $expected = $this->sampleBase64String();
@@ -212,9 +191,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $result);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_generate_base64_from_cache()
     {
         $cachedAvatar = 'data:image/png;base64,iVBO';
@@ -228,9 +205,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($cachedAvatar, $result);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_generate_file()
     {
         $file = __DIR__.'/avatar.png';
@@ -243,9 +218,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         unlink($file);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_generate_circle_svg()
     {
         $expected = '<svg xmlns="http://www.w3.org/2000/svg" width="100" height="100" viewBox="0 0 100 100">';
@@ -266,9 +239,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $svg);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_generate_rectangle_svg()
     {
         $expected = '<svg xmlns="http://www.w3.org/2000/svg" width="100" height="100" viewBox="0 0 100 100">';
@@ -290,9 +261,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $svg);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_generate_responsive_svg()
     {
         $expected = '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100">';
@@ -315,9 +284,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $svg);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_generate_svg_with_custom_font_family()
     {
         $expected = '<svg xmlns="http://www.w3.org/2000/svg" width="100" height="100" viewBox="0 0 100 100">';
@@ -339,9 +306,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $svg);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_use_the_foreground_color_for_the_svg_border()
     {
         $expected = '<svg xmlns="http://www.w3.org/2000/svg" width="100" height="100" viewBox="0 0 100 100">';
@@ -361,9 +326,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $svg);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_use_the_background_color_for_the_svg_border()
     {
         $expected = '<svg xmlns="http://www.w3.org/2000/svg" width="100" height="100" viewBox="0 0 100 100">';
@@ -383,9 +346,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $svg);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_handle_html_entities()
     {
         $expected = '<svg xmlns="http://www.w3.org/2000/svg" width="100" height="100" viewBox="0 0 100 100">';
@@ -406,9 +367,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $svg);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_generate_gravatar()
     {
         $expected = 'https://www.gravatar.com/avatar/db6f5ab11fb203026beb0e298930cc5a07080022e7cbb4c597b97321585df61b?s=88';
@@ -422,9 +381,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $url);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_generate_gravatar_with_size()
     {
         $expected = 'https://www.gravatar.com/avatar/db6f5ab11fb203026beb0e298930cc5a07080022e7cbb4c597b97321585df61b?s=100';
@@ -437,9 +394,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $url);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_generate_gravatar_with_default()
     {
         $expected = 'https://www.gravatar.com/avatar/db6f5ab11fb203026beb0e298930cc5a07080022e7cbb4c597b97321585df61b?d=identicon&s=100';
@@ -452,9 +407,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $url);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_generate_gravatar_with_default_and_rating()
     {
         $expected = 'https://www.gravatar.com/avatar/db6f5ab11fb203026beb0e298930cc5a07080022e7cbb4c597b97321585df61b?d=identicon&r=pg&s=100';
@@ -467,9 +420,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $url);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_generate_gravatar_with_size_overriden()
     {
         $expected = 'https://www.gravatar.com/avatar/db6f5ab11fb203026beb0e298930cc5a07080022e7cbb4c597b97321585df61b?s=300';
@@ -482,9 +433,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $url);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_set_background()
     {
         $hex = '#ffffff';
@@ -495,9 +444,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($hex, $avatar->getAttribute('background'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_set_foreground()
     {
         $hex = '#ffffff';
@@ -508,9 +455,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($hex, $avatar->getAttribute('foreground'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_set_dimension()
     {
         $avatar = new \Laravolt\Avatar\Avatar();
@@ -524,9 +469,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals(10, $avatar->getAttribute('height'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_set_font_size()
     {
         $size = 12;
@@ -537,9 +480,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($size, $avatar->getAttribute('fontSize'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_set_font_family()
     {
         $font = 'Lato';
@@ -550,9 +491,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($font, $avatar->getAttribute('fontFamily'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_set_border()
     {
         $borderSize = 1;
@@ -568,9 +507,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         }
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_set_border_radius()
     {
         $borderSize = 1;
@@ -586,9 +523,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         }
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_accept_valid_shape()
     {
         $shapes = ['circle', 'square'];
@@ -601,9 +536,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         }
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_throw_exception_for_invalid_shape()
     {
         $this->expectException(InvalidArgumentException::class);
@@ -613,9 +546,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $avatar->setShape($shape)->buildAvatar();
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_get_raw_image_object()
     {
         $avatar = new \Laravolt\Avatar\Avatar();
@@ -624,9 +555,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertInstanceOf(\Intervention\Image\Image::class, $imageObject);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_get_initial()
     {
         $avatar = new \Laravolt\Avatar\Avatar();
@@ -635,9 +564,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals('CK', $avatar->getInitial());
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_cast_to_string()
     {
         $expected = $this->sampleBase64String();
@@ -647,9 +574,7 @@ class AvatarPhpTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($expected, $result);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_can_set_custom_generator()
     {
         $avatar = new \Laravolt\Avatar\Avatar();
